P4/web/requirements.txt
flask
flask_wtf
pyjwt
redis
requests
python-dotenv
authlib
six
flask_talisman